happifyml/templates/pl_research_template/requirements.txt
# --------- pytorch --------- #
torch==1.8.1
torchvision==0.9.1
pytorch-lightning>=1.3.2
torchmetrics>=0.3.2

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0
# hydra-ax-sweeper==1.1.0
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.30
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- transformers --------- #
transformers>=4.6.1
datasets==1.8.0
accelerate==0.3.0
seqeval==1.2.2

# --------- linters --------- #
pre-commit
black
isort
flake8

# --------- others --------- #
jupyterlab
python-dotenv
rich
pytest
sh
pudb
# dvc
scikit-learn>=0.24.2
seaborn>=0.11.1
pandas>=1.2.4
pyyaml>=5.4.1
spacy==3.1.0
py3nvml>=0.2.6
